Overhaul addding / removing packages; Add client profile
The client now gets information about the actual changes that have been
made on the coordinator. Also made some effort to replace instances
where we use a list of packages with a HashSet. There is no reason to
ever name a package more than once, so this should prevent any
tomfoolery from occuring. Also this allows the usage of set comparison
which is quite nice.

Also added the ability for the client to have multiple config profiles.
This helps with working with multiple coordinator instances during
development.